+++ src/seqs/index.ts
@@ -1,0 +1,3 @@
+export * from './seq';
+export * from './seqs';
+export * from './spec';
